discord_share/requirements.txt
discord
python-dotenv
pymongo
pymongo[srv]

